Move HmdaFileParser events to persistenceModel project
This will allow us to use those events in persistence model
serialization. Also moves LarParsingError object to model project, where
it can be included as a dependency in persistenceModel and parser
projects.
