+++ api/src/main/scala/hmda/api/model/ParsingErrorSummary.scala
@@ -1,7 +1,7 @@
 package hmda.api.model
 
 import hmda.model.fi.SubmissionStatus
-import hmda.parser.fi.lar.LarParsingError
+import hmda.model.parser.LarParsingError
 
 case class ParsingErrorSummary(
   transmittalSheetErrors: Seq[String],
