+++ model/shared/src/main/scala/hmda/model/parser/LarParsingError.scala
@@ -1,3 +1,3 @@
-package hmda.parser.fi.lar
+package hmda.model.parser
 
 case class LarParsingError(lineNumber: Int = 0, errorMessages: List[String] = List.empty)
