+++ persistence/src/test/scala/hmda/persistence/processing/HmdaFileParserSpec.scala
@@ -7,9 +7,10 @@ import akka.testkit.TestProbe
 import org.scalatest.BeforeAndAfterEach
 import com.typesafe.config.ConfigFactory
 import hmda.model.fi.SubmissionId
-import hmda.parser.fi.lar.LarParsingError
+import hmda.model.parser.LarParsingError
 import hmda.parser.fi.ts.TsCsvParser
 import hmda.persistence.messages.CommonMessages.GetState
+import hmda.persistence.messages.events.processing.HmdaFileParserEvents.{ LarParsedErrors, TsParsed, TsParsedErrors }
 import hmda.persistence.model.ActorSpec
 import hmda.persistence.processing.HmdaFileParser._
 import hmda.persistence.processing.HmdaRawFile._
